keras_mnist/requirements.txt
seldon-core
h5py==2.10.0
keras==2.4.3
tensorflow==2.3.1
numpy==1.17.4
Pillow==7.1.0